Updtating the iOS sources for iOS 8.

+++ var/apps/iphone/Siberian/SiberianCMS/Application/Cache/EVURLCache.m
@@ -181,7 +181,7 @@ static NSString* _preCacheDirectory;
         return result == 0;
     } else {
         // First try and remove the extended attribute if it is present
-        int result = getxattr(filePath, attrName, NULL, sizeof(u_int8_t), 0, 0);
+        long result = getxattr(filePath, attrName, NULL, sizeof(u_int8_t), 0, 0);
         if (result != -1) {
             // The attribute exists, we need to remove it
             int removeResult = removexattr(filePath, attrName, 0);
